astro_airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
snowflake-connector-python
polars
pandas
requests
python-dotenv
pyarrow